cleaner compoent arch
